+++ src/api/ComputeEngineAPI.java
@@ -1,7 +1,5 @@
 package api;
 
-import project.annotations.ConceptualAPI;
-
 @ConceptualAPI
 public interface ComputeEngineAPI {
     int computeFactorial(int number);
