+++ src/api/ConceptualAPIPrototype.java
@@ -1,4 +1,4 @@
-package project.annotations;
+package api;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,6 +7,6 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface ProcessAPIPrototype {
-	// Marker annotation, should be applied to a method within a prototype class
+public @interface ConceptualAPIPrototype {
+  // Marker annotation, should be applied to a method within a prototype class
 }
