+++ src/api/NetworkAPI.java
@@ -1,12 +1,12 @@
-package project.annotations;
+package api;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.METHOD)
+@Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface NetworkAPIPrototype {
-	// Marker annotation, should be applied to a method within a prototype class
+public @interface NetworkAPI {
+  // Marker annotation, should be applied to an interface type
 }
