+++ src/api/NetworkAPIPrototype.java
@@ -1,12 +1,12 @@
-package project.annotations;
+package api;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.TYPE)
+@Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface ProcessAPI {
-	// Marker annotation, should be applied to an interface type
+public @interface NetworkAPIPrototype {
+	// Marker annotation, should be applied to a method within a prototype class
 }
