+++ src/api/ProcessAPI.java
@@ -1,4 +1,4 @@
-package project.annotations;
+package api;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,6 +7,6 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface ConceptualAPI {
+public @interface ProcessAPI {
 	// Marker annotation, should be applied to an interface type
 }
