+++ src/api/StorageComputeAPI.java
@@ -1,7 +1,5 @@
 package api;
 
-import project.annotations.ProcessAPI;
-
 @ProcessAPI
 public interface StorageComputeAPI {
     int[] readNumbers(String source);
